+++ tikapy/__init__.py
@@ -225,16 +225,18 @@ class TikapySslClient(TikapyBaseClient):
     RouterOS SSL API Client.
     """
 
-    def __init__(self, address, port=8729):
+    def __init__(self, address, port=8729, verify_cert=True):
         """
         Initialize client.
 
         :param address: Remote device address (maybe a hostname)
         :param port: Remote device port (defaults to 8728)
+        :param verify_cert: Verify device certificate against system CAs
         """
         super().__init__()
         self.address = address
         self.port = port
+        self.verify_cert = verify_cert
 
     def _connect(self):
         """
@@ -242,7 +244,10 @@ class TikapySslClient(TikapyBaseClient):
         """
         self._connect_socket()
         try:
-            self._sock = ssl.wrap_socket(self._base_sock)
+            ctx = ssl.create_default_context()
+            if not self.verify_cert:
+                ctx.verify_mode = ssl.CERT_OPTIONAL
+            self._sock = ctx.wrap_socket(self._base_sock)
         except ssl.SSLError:
             LOG.error('could not establish SSL connection')
             raise ClientError('could not establish SSL connection')